Refactor: Adjust placement of inlined utility functions
Moved inlined helper functions in `use-dismiss.ts` and `use-floating.ts`
to dedicated sections at the end of each file to better align with
existing code conventions.

Changes:
- In `src/composables/interactions/use-dismiss.ts`:
    - Functions `normalizeProp`, `isHTMLElement`, `isEventTargetWithin`, `isClickOnScrollbar`, and `useComposition` moved to a new "Helper Functions" section at the end of the file.
- In `src/composables/use-floating.ts`:
    - Functions `roundByDPR` and `getDPR` moved to the existing "Utility Functions" section at the end of the file.

This improves consistency in code structure within these files.
