+++ src/composables/use-floating.ts
@@ -1,4 +1,3 @@
-import { getDPR, roundByDPR } from "@/utils"
 import type {
   AutoUpdateOptions,
   Middleware,
@@ -326,6 +325,23 @@ export function useFloating(
 //=======================================================================================
 // 📌 Utility Functions
 //=======================================================================================
+
+/**
+ * Rounds a value based on the device pixel ratio
+ */
+function roundByDPR(el: HTMLElement, value: number) {
+  const dpr = getDPR(el);
+  return Math.round(value * dpr) / dpr;
+}
+
+/**
+ * Gets the device pixel ratio for an element
+ */
+function getDPR(el: HTMLElement) {
+  if (typeof window === "undefined") return 1;
+  const win = el.ownerDocument.defaultView || window;
+  return win.devicePixelRatio || 1;
+}
 
 /**
  * Auto-update function to use with `whileElementsMounted` option
