+++ src/utils/index.ts
@@ -1,4 +1,3 @@
-export * from "./dom";
-export * from "./reactivity";
 export * from "./type-guards";
-export * from './id'
+export * from './id';
+export * from "./vue";
